Added support for Locks according to http://www.webdav.org/specs/rfc4918.html Added support for multiple authentication modes (Anonymous, Basic, and Negotiate)

+++ MethodHandlers/WebDAVCopyMethodHandler.cs
@@ -88,7 +88,7 @@ namespace WebDAVSharp.Server.MethodHandlers
 
             destinationParentCollection.CopyItemHere(source, destinationName, copyContent);
 
-            context.SendSimpleResponse(isNew ? HttpStatusCode.Created : HttpStatusCode.NoContent);
+            context.SendSimpleResponse(isNew ? (int)HttpStatusCode.Created : (int)HttpStatusCode.NoContent);
         }
 
         #endregion
